+++ frontend/src/components/Horoscope/Tarot.jsx
@@ -1,13 +1,18 @@
 import React, { useState, useEffect } from "react";
+import { Modal, Button } from "react-bootstrap";
 import "../../styles/TarotAnimation.css";
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
 
 function Intro({ onStart }) {
   const [selectedCategory, setSelectedCategory] = useState(null);
   const [isShuffling, setIsShuffling] = useState(false);
-  const [step, setStep] = useState("gather"); // 'gather' | 'swap'
+  const [step, setStep] = useState("gather");
   const [selectedIndex, setSelectedIndex] = useState(null);
   const [shuffleImageIndex, setShuffleImageIndex] = useState(null);
+
+  // ✅ 모달 상태
+  const [modalMessage, setModalMessage] = useState("");
+  const [showModal, setShowModal] = useState(false);
 
   const getCategoryName = (id) => {
     switch (id) {
@@ -31,43 +36,45 @@ function Intro({ onStart }) {
 
   const handleCardClick = (index) => {
     setSelectedIndex(index);
-    let categoryId = index + 1; // 0 -> 1, 1 -> 2, 2 -> 3
-    setSelectedCategory(categoryId);
+    setSelectedCategory(index + 1);
   };
 
   const handleStart = async () => {
     const token = localStorage.getItem("token");
-  
+
     if (!token) {
-      alert("로그인이 필요합니다. 로그인 후 이용해주세요!");
+      setModalMessage("로그인이 필요합니다. 로그인 후 이용해주세요!");
+      setShowModal(true);
       return;
     }
-  
+
     if (selectedIndex === null) {
-      alert("카드를 먼저 선택해주세요!");
+      setModalMessage("카드를 먼저 선택해주세요!");
+      setShowModal(true);
       return;
     }
-  
+
     try {
       const res = await fetch(`${API_BASE_URL}/api/tarot/check`, {
         headers: { Authorization: `Bearer ${token}` }
       });
-  
+
       const data = await res.json();
       if (data.played) {
-        alert("오늘의 타로는 이미 진행하셨습니다. 내일 다시 시도해주세요!");
+        setModalMessage("오늘의 타로는 이미 진행하셨습니다. 내일 다시 시도해주세요!");
+        setShowModal(true);
         return;
       }
-  
+
       setShuffleImageIndex(selectedIndex);
       setIsShuffling(true);
       setStep("gather");
     } catch (err) {
       console.error("타로 체크 실패:", err);
-      alert("서버 오류로 타로를 시작할 수 없습니다.");
+      setModalMessage("서버 오류로 타로를 시작할 수 없습니다.");
+      setShowModal(true);
     }
   };
-  
 
   const cardClass = `intro-cards ${isShuffling ? "shuffle-sequence" : ""} ${step === "swap" ? "step2" : ""}`;
 
@@ -110,6 +117,18 @@ function Intro({ onStart }) {
           {selectedCategory ? `${getCategoryName(selectedCategory)} 타로 Start` : "카드를 선택하세요"}
         </button>
       )}
+
+      {/* ✅ 알림 모달 */}
+      <Modal show={showModal} onHide={() => setShowModal(false)} centered>
+        <Modal.Body className="text-center py-4">
+          <p style={{ fontSize: "1.1rem" }}>{modalMessage}</p>
+          <div className="mt-3">
+            <Button variant="primary" onClick={() => setShowModal(false)}>
+              확인
+            </Button>
+          </div>
+        </Modal.Body>
+      </Modal>
     </div>
   );
 }
