+++ frontend/src/components/PlaceDetail/ReportModal.jsx
@@ -1,4 +1,3 @@
-// src/components/PlaceDetail/ReportModal.jsx
 import React, { useState } from "react";
 import { Modal, ListGroup, Form, Button } from "react-bootstrap";
 
@@ -10,9 +9,14 @@ function ReportModal({ show, onHide, onSelect, type = "opinion" }) {
   const [selectedReason, setSelectedReason] = useState(null);
   const [customReason, setCustomReason] = useState("");
 
+  // ✅ 모달 메시지용 상태
+  const [modalMessage, setModalMessage] = useState("");
+  const [showMessageModal, setShowMessageModal] = useState(false);
+
   const handleSubmit = () => {
     if (selectedReason === "기타" && !customReason.trim()) {
-      alert("사유를 입력해주세요");
+      setModalMessage("사유를 입력해주세요.");
+      setShowMessageModal(true);
       return;
     }
     const finalReason = selectedReason === "기타" ? customReason : selectedReason;
@@ -28,42 +32,56 @@ function ReportModal({ show, onHide, onSelect, type = "opinion" }) {
   };
 
   return (
-    <Modal show={show} onHide={handleCancel} centered>
-      <Modal.Header closeButton>
-        <Modal.Title>신고 사유 선택</Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
-        <ListGroup>
-          {reasons.map((reason) => (
-            <ListGroup.Item
-              action
-              active={selectedReason === reason}
-              key={reason}
-              onClick={() => setSelectedReason(reason)}
-            >
-              {reason}
-            </ListGroup.Item>
-          ))}
-        </ListGroup>
+    <>
+      <Modal show={show} onHide={handleCancel} centered>
+        <Modal.Header closeButton>
+          <Modal.Title>신고 사유 선택</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <ListGroup>
+            {reasons.map((reason) => (
+              <ListGroup.Item
+                action
+                active={selectedReason === reason}
+                key={reason}
+                onClick={() => setSelectedReason(reason)}
+              >
+                {reason}
+              </ListGroup.Item>
+            ))}
+          </ListGroup>
 
-        {selectedReason === "기타" && (
-          <Form.Control
-            className="mt-3"
-            placeholder="사유를 입력해주세요"
-            value={customReason}
-            onChange={(e) => setCustomReason(e.target.value)}
-          />
-        )}
-      </Modal.Body>
-      <Modal.Footer>
-      <Button variant="danger" onClick={handleSubmit} disabled={!selectedReason}>
-          신고 제출
-        </Button>
-        <Button variant="secondary" onClick={handleCancel}>
-          취소
-        </Button>
-      </Modal.Footer>
-    </Modal>
+          {selectedReason === "기타" && (
+            <Form.Control
+              className="mt-3"
+              placeholder="사유를 입력해주세요"
+              value={customReason}
+              onChange={(e) => setCustomReason(e.target.value)}
+            />
+          )}
+        </Modal.Body>
+        <Modal.Footer>
+          <Button variant="danger" onClick={handleSubmit} disabled={!selectedReason}>
+            신고 제출
+          </Button>
+          <Button variant="secondary" onClick={handleCancel}>
+            취소
+          </Button>
+        </Modal.Footer>
+      </Modal>
+
+      {/* ❗입력 누락 시 경고용 모달 */}
+      <Modal show={showMessageModal} onHide={() => setShowMessageModal(false)} centered>
+        <Modal.Body className="text-center py-4">
+          <p style={{ fontSize: "1.1rem", margin: 0 }}>{modalMessage}</p>
+          <div className="mt-3">
+            <Button variant="primary" onClick={() => setShowMessageModal(false)}>
+              확인
+            </Button>
+          </div>
+        </Modal.Body>
+      </Modal>
+    </>
   );
 }
