+++ frontend/src/pages/MyPage.jsx
@@ -13,8 +13,7 @@ import OpinionTab from "../components/MyPage/OpinionTab";
 import ReportTab from "../components/MyPage/ReportTab";
 import BookmarkTab from "../components/MyPage/BookmarkTab";
 import EditUserInfo from "../components/MyPage/EditUserInfo";
-import AlarmTab from "../components/MyPage/AlarmTab";
-import AlarmListTab from "../components/MyPage/AlarmListTab";
+import AlarmManagerTab from "../components/MyPage/AlarmManagerTab";
 import { FaBell } from "react-icons/fa";
 import { FaUser, FaCommentDots, FaExclamationCircle, FaHatWizard } from "react-icons/fa";
 import "../styles/Mypage.css";
@@ -134,11 +133,8 @@ function MyPage() {
             {activeTab === "opinions" && <OpinionTab userInfo={userInfo} />}
             {activeTab === "reports" && <ReportTab userInfo={userInfo} />}
             {activeTab === "tarot" && <TodayTarotResultTab userId={userInfo?.userId} />}
-            {activeTab === "alarms" &&   <>
-              <AlarmTab userInfo={userInfo} />
-              <AlarmListTab userId={userInfo?.userId} />
-            </>
-            }
+            {activeTab === "alarms" && <AlarmManagerTab userInfo={userInfo} />}
+
           </Col>
         </Row>
       </div>
