+++ frontend/src/styles/TodayPlace/PlaceDetail.css
@@ -240,3 +240,30 @@
   overflow: hidden;
   white-space: nowrap;
 }
+.info-block {
+  margin-bottom: 1.5rem;
+  font-size: 0.95rem;
+}
+
+.info-row {
+  display: flex;
+  margin-bottom: 0.4rem;
+}
+
+.info-label {
+  width: 120px;
+  font-weight: bold;
+  color: #444;
+}
+
+.info-value {
+  flex: 1;
+  color: #222;
+}
+.rating-under-title {
+  font-size: 1rem;
+  color: #333;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
